+++ packages/rest/src/index.ts
@@ -1,2 +1,3 @@
 export * from './rest';
 export * from './interceptors';
+export * from './lib';