+++ packages/rest/src/rest/rest.ts
@@ -1,10 +1,6 @@
-import {
-  ApiResponse,
-  BodyParserType,
-  EnhancedRequestOptions,
-  RestClientConfiguration,
-} from '@ts-fetcher/types';
-import { chainRequestInterceptors, chainResponseInterceptors } from '../';
+import { ApiResponse, EnhancedRequestOptions, RestClientConfiguration } from '@ts-fetcher/types';
+import { chainRequestInterceptors, chainResponseInterceptors, sleepWithCallback } from '../';
+import { types } from 'node:util';
 
 export class Rest {
   public origin: string;
@@ -73,11 +69,26 @@ export class Rest {
   public async request<RES = unknown, REQ = unknown>(
     options: EnhancedRequestOptions<REQ>
   ): Promise<ApiResponse<RES>> {
+    //===============LIFECYCLE================
+    if (this.restOptions.defaultRequestOptions) {
+      options = { ...this.restOptions.defaultRequestOptions, ...options };
+    }
     const interceptors = this.mergeInterceptors(options);
     if (interceptors.request?.length) {
       options = chainRequestInterceptors(options, interceptors.request);
     }
+    if (options.lifecycle?.onRequestInit) {
+      const requestInit = options.lifecycle.onRequestInit(options, this.restOptions);
+      if (requestInit.forceReturn) {
+        this.makeRequest(options);
+        return this.responseWithInterceptors(
+          this.makeResponse(requestInit.data, options, false, true),
+          interceptors
+        );
+      }
+    }
 
+    // ==================CACHE RESPONSE==================
     if (!this.restOptions.cache && options.cache) {
       throw new Error('Rest cache options is not provided!');
     }
@@ -94,23 +105,41 @@ export class Rest {
       }
     }
 
-    const body =
-      options.body && options.method !== 'GET'
-        ? this.parseBody(options.body.parseAs, options.body.data)
-        : null;
+    //=================REFETCH===================
+    let response = await this.makeRequest(options);
 
-    const response = await fetch(`${options.origin ?? this.origin}/${options.path}`, {
-      // @ts-expect-error I hate ts
-      body,
-      ...options,
-    });
+    if (!response.ok && options.refetch) {
+      if (
+        options.refetch.delay <= 0 ||
+        !Number.isFinite(options.refetch.delay) ||
+        !Number.isInteger(options.refetch.delay)
+      ) {
+        throw new Error('Refetch delay must be a valid number');
+      }
+      for (let i = 0; i < options.refetch.attemps; i++) {
+        const isSuccess = await sleepWithCallback(options.refetch.delay, async () => {
+          const res = await this.makeRequest(options);
+          response = res;
+          return res.ok;
 
-    const formated = {
-      success: response.ok,
-      data: await this.parseJsonResponse<RES>(response),
-      // TODO: cache logic
-      cached: true,
-    };
+        });
+        if (isSuccess) {
+          break;
+        }
+
+        if (!isSuccess && options.refetch.attemps === i + 1) {
+          throw new Error(`Refetch attemps failed...\n ${response}`);
+        }
+      }
+    }
+
+    //================REQUEST END================
+    const formated = this.makeResponse(
+      await this.parseJsonResponse(response),
+      options,
+      true,
+      response.ok
+    );
 
     if (options.cache && this.restOptions.cache) {
       await this.restOptions.cache.set(
@@ -127,6 +156,33 @@ export class Rest {
       },
       interceptors
     );
+  }
+
+  public async makeRequest<REQ extends EnhancedRequestOptions = EnhancedRequestOptions>(
+    options: REQ
+  ) {
+    const body = options.body && options.method !== 'GET' ? this.parseBody(options.body) : null;
+
+    // @ts-expect-error I hate ts
+    return await fetch(`${options.origin ?? this.origin}${options.path}`, {
+      body,
+      ...options,
+    });
+  }
+
+  public makeResponse<REQ extends EnhancedRequestOptions = EnhancedRequestOptions>(
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    data: any,
+    options: REQ,
+    cached = false,
+    success = true
+  ): ApiResponse<typeof data, REQ, typeof options.method> {
+    return {
+      data,
+      cached,
+      success,
+      options,
+    };
   }
 
   public responseWithInterceptors<RES = unknown>(
@@ -156,20 +212,49 @@ export class Rest {
   }
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  public parseBody(parseAs: BodyParserType, data: any) {
-    switch (parseAs) {
-      case 'JSON':
-        return JSON.stringify(data);
-      default:
-        return new FormData(data);
+  public async parseBody(body: any, requestBody = true) {
+    if (body == null) {
+      return null;
+    } else if (typeof body === 'string') {
+      try {
+        return requestBody ? body : JSON.parse(body);
+      } catch {
+        return body;
+      }
+    } else if (types.isUint8Array(body)) {
+      return body;
+    } else if (types.isArrayBuffer(body)) {
+      return new Uint8Array(body);
+    } else if (body instanceof URLSearchParams) {
+      return body.toString();
+    } else if (body instanceof DataView) {
+      return new Uint8Array(body.buffer);
+    } else if (body instanceof Blob) {
+      return new Uint8Array(await body.arrayBuffer());
+    } else if (body instanceof FormData) {
+      return body;
+    } else if ((body as Iterable<Uint8Array>)[Symbol.iterator]) {
+      const chunks = [...(body as Iterable<Uint8Array>)];
+
+      return Buffer.concat(chunks);
+    } else if ((body as AsyncIterable<Uint8Array>)[Symbol.asyncIterator]) {
+      const chunks: Uint8Array[] = [];
+
+      for await (const chunk of body as AsyncIterable<Uint8Array>) {
+        chunks.push(chunk);
+      }
+
+      return Buffer.concat(chunks);
     }
+
+    throw new TypeError(`Unable to resolve body.`);
   }
 
-  public async parseJsonResponse<RES = unknown>(res: Response) {
+  public async parseJsonResponse(res: Response) {
     try {
-      return (await res.json()) as RES;
+      return await res.json();
     } catch {
-      return null as RES;
+      return null;
     }
   }
 }