+++ packages/rest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ts-fetcher/rest
 
+## 1.2.0
+
+### Minor Changes
+
+- Implemented refetch system and global options <br>
+  Remove createCache factory <br>
+  Simplified request bodies
+
 ## 1.1.5
 
 ### Patch Changes